requirements.txt
Flask==3.0.3
flask-cors==4.0.1
bcrypt==4.1.3
huggingface_hub
python-dotenv==1.0.1
requests==2.32.3
PyMySQL==1.1.1
twilio==9.3.6
cryptography==42.0.8